#69 Did some work on the settings page
