+++ app/templates/admin/index.html
@@ -3,5 +3,8 @@
 {% block body %}
 
     <h1>Admin Dashboard</h1>
+
+    <h3>Welcome</h2>
+    <p>Add something here</p>
     
 {% endblock %}
