+++ app/templates/chat/chat1.html
@@ -84,7 +84,7 @@
     const myUsername = "{{ username }}";
 
     const socket = io(
-        "http://10.153.202.211:5000", {
+        "http://192.168.68.64:5000", { // FIXME: Automatically change this to the server's IP address
             auth: {
                 token: "123"
             }
