+++ app/views/views.py
@@ -1,5 +1,5 @@
 from app import app, db
-from flask import render_template, request, url_for, redirect
+from flask import render_template
 from flask_login import current_user
 
 @app.route("/")
@@ -13,19 +13,6 @@ def home():
     if current_user.is_authenticated:
         return render_template("home.html", user=current_user)
     return render_template("home.html", user=None)
-
-@app.route("/update_image_link", methods=['POST'])
-def update_image_link():
-    if current_user.is_authenticated:
-        image_link = request.form.get('image_link')
-
-        # You may want to add validation logic here to ensure the link is valid
-
-        # Update the user's image_link in the database
-        current_user.image_link = image_link
-        db.session.commit()
-
-    return redirect(url_for('home'))
 
 
 @app.route("/new_home")
